EventLog options
MSCT has the ability to write out the SoliNet or _extended_ style
time tags: this is enabled with the MSCT_EVLOG_EXTENDED
environment variable.

SoliNet has the ability to add the thread id to logging statement
indicating which thread wrote the log statement: this is enabled
with the EVENTLOG_DEBUG_THREAD_ID environment variable.