Replacement grep program if available
The check is pretty cool as well and something I should probably
make greater use of in the future.  The functionality is to look
for the ag executable (which is supposed to be a better awk which
itself is supposed to be a better grep for programmers) and if it
is present then use that as vim's grep program.