Adding ctrlp vim plugin
People have been raving about this online for a long while so
I figured it was time to try it out.  Have not used it yet but I
do get annoyed at the file navigation part of the vim workflow I
have especially with getting to Soli{Net,Math} from MSCT, etc.